Merge: Refactor to norm and general merge fixes
